Created package for localization, refactored code in MainController
